// src/Components/UI/Button/enums.ts
export enum ButtonSizes {
  small = 'size-s',
  smallMedium = 'size-sm',
  medium = 'size-m',
  extraMedium = 'size-m-2',
  extraLarge = 'size-xl'
}

export enum ButtonColors {
  white = 'color-white',
  black = 'color-black',
  purple = 'color-purple'
}

export enum ButtonBackgrounds {
  white = 'background-white',
  black = 'background-black',
  orange = 'background-orange',
  green = 'background-green',
  orangeBold = 'background-orange-bold',
  blue = 'background-blue'
}

export enum ButtonTypes {
  filled = 'type-filled',
  outlined = 'type-outlined',
  borderless = 'type-borderless'
}

export enum ButtonBorderRadius {
  br5 = 'border-radius-5',
  br10 = 'border-radius-10',
  br15 = 'border-radius-15',
  br20 = 'border-radius-20',
  br100 = 'border-radius-100'
}

export enum ButtonSpinnerColors {
  white = 'spinner-color-white',
  black = 'spinner-color-black',
  purple = 'spinner-color-purple'
}

export enum MarketPlaces {
  PUBLIC = 'PUBLIC',
  NRC_SUDAN = 'NRC_SUDAN',
  MYANMAR = 'MYANMAR',
  MEXICO = 'MEXICO'
}